+++ src/lib/index.js
@@ -27,3 +27,5 @@ export { default as Headline } from "./Headline/Headline.svelte";
 export { default as PymChild } from "./Pym/PymChild.svelte";
 export { default as LoadingWrapper } from "./LoadingWrapper/LoadingWrapper.svelte";
 export { default as Tooltip } from "./Tooltip/Tooltip.svelte";
+export { default as AxisY } from "./Layercake/AxisY.svelte";
+export { default as AxisX } from "./Layercake/AxisX.svelte";